Refactor preload and authButton
