+++ desktop/src/preload/api/index.ts
@@ -1,5 +1,2 @@
 export * from "./checkAPIStatus";
 export * from "./checkIPRange";
-export * from "./getUser";
-export * from "./safePromise";
-export * from "./fetchProtectedRoute";
