+++ desktop/src/preload/api/sso/getUser.ts
