+++ desktop/src/preload/api/sso/index.ts
@@ -1,0 +1,2 @@
+export * from "./fetchProtectedRoute";
+export * from "./getUser";
