+++ desktop/src/preload/api/utils/safePromise.ts
