+++ desktop/src/preload/index.d.ts
@@ -9,15 +9,21 @@ declare global {
 			checkApiStatus: (url: string) => Promise<boolean>;
 			checkIpRange: () => Promise<boolean>;
 			getCurrentApiUrl: () => Promise<string>;
-			startLoginProcess: () => Promise<void>;
-			logout: (idToken: string | undefined) => Promise<void>;
-			getUser: (accessToken: string | undefined) => SSOUser<IdirIdentityProvider> | undefined;
-			safePromise: <T>(promise: Promise<T>) => Promise<[Error | null, T | null]>;
-			fetchProtectedRoute: (
-				url: string,
-				accessToken: string | undefined,
-				options: RequestInit = {},
-			) => Promise<[Error | null, Response | null]>;
+
+			sso: {
+				startLoginProcess: () => Promise<void>;
+				logout: (idToken: string | undefined) => Promise<void>;
+				getUser: (accessToken: string | undefined) => SSOUser<IdirIdentityProvider> | undefined;
+				fetchProtectedRoute: (
+					url: string,
+					accessToken: string | undefined,
+					options: RequestInit = {},
+				) => Promise<[Error | null, Response | null]>;
+			};
+
+			utils: {
+				safePromise: <T>(promise: Promise<T>) => Promise<[Error | null, T | null]>;
+			};
 		};
 	}
 }
