+++ desktop/src/renderer/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from "./VPNPopup";
 export * from "./Versions";
+export * from "./AuthButton";
